contract-reply-app/backend/requirements.txt
flask
flask-cors
openai
python-dotenv
PyPDF2
python-docx
werkzeug
